# config/swaggerUser.json
{
    "swagger": "2.0",
    "info": {
        "title": "CSF Healthcare",
        "description": "CSF Healthcare api documentation for users",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "192.168.102.76:3002",
    "basePath": "",
    "paths": {
        "/auth-api/testEncryption": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "testEncryption",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "facebookId",
                        "in": "formData",
                        "description": "facebookId is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "role",
                        "in": "formData",
                        "description": "role is required [USER/PROFESSIONAL]",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/registerUser": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "registerUser",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "Registration step number 1",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "description": "Full Name",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email (Optional)",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "required": true,
                        "description": "Phone number is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "password",
                        "required": true,
                        "in": "formData",
                        "description": "Password is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "countryCode is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "role",
                        "in": "formData",
                        "description": "role is required - [ USER/PROFESSIONAL/FACILITY]",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/verifyOtp": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "verifyOtp",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "userId",
                        "in": "formData",
                        "description": "UserId is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "otp",
                        "in": "formData",
                        "description": "otp is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "phone is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "merge",
                        "in": "formData",
                        "description": "(OPTIONAL)",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "facebookId",
                        "in": "formData",
                        "description": "(OPTIONAL))",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/userLogin": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "userLogin",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "phoneNumber is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "countryCode is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/socialLogin": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "socialLogin",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "facebookId",
                        "in": "formData",
                        "description": "facebookId is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "(OPTIONAL)",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "(OPTIONAL if user is logging in) (REQUIRED if user is getting registered)",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "(OPTIONAL if user is logging in) (REQUIRED if user is getting registered)",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "role",
                        "in": "formData",
                        "description": "(OPTIONAL if user is logging in) (REQUIRED if user is getting registered)",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/socialLoginMerge": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "socialLoginMerge",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "merge by phone number",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "userId",
                        "in": "formData",
                        "description": "userId is required (userId received from socialLogin API)",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "countryCode is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "phone is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "facebookId",
                        "in": "formData",
                        "description": "facebookId is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "role",
                        "in": "formData",
                        "description": "role is required [USER/PROFESSIONAL/FACILITY]",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/facebookLoginMerge": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "facebookLoginMerge",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "merge by facebook id",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "facebookId",
                        "in": "formData",
                        "description": "facebookId is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "role",
                        "in": "formData",
                        "description": "role is required [USER/PROFESSIONAL]",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/resendOtp": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "resendOtp",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "Country Code is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Phone is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/forgotPassword": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "forgotPassword",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "Country Code is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Phone is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/updatePassword": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updatePassword",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "phone is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "countryCode",
                        "in": "formData",
                        "description": "countryCode is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "otp",
                        "in": "formData",
                        "description": "otp is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/auth-api/appVersioning": {
            "get": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "appVersioning",
                "tags": [
                    "1 User Auth APIs"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },{
                        "name": "deviceType",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },{
                        "name": "appVersion",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
...
    }
}